requirements.txt
requests==2.32.3
pypdf==5.4.0
openai== 1.68.2
pymupdf==1.24.0
python-dotenv==1.0.0
minio==7.2.3
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
